scripts/scraper/requirements.txt
# Web Scraper Dependencies
# Install with: pip install -r requirements.txt

# HTTP requests and caching
requests>=2.31.0
requests-cache>=1.1.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Fuzzy matching (faster than fuzzywuzzy)
rapidfuzz>=3.5.0

# Supabase client
supabase>=2.0.0

# Environment variables
python-dotenv>=1.0.0

# URL parsing and sitemap handling
urllib3>=2.0.0
xml.etree.ElementTree  # Built-in, no install needed
